requirements.txt
streamlit
pandas
requests
PyYAML
rapidfuzz
sshtunnel
SQLAlchemy
PyMySQL
sshtunnel==0.4.0
paramiko==2.12.0
python-dotenv
st-copy-to-clipboard
streamlit_extras
streamlit_nested_layout

